Failing test suite for `Hastory.Cli.OptParseSpec`.
The test suite - in particulary `Hastory.Cli.OptParseSpec` - is
currently failing. The reason for this is the `gather` command is
currently treating `RemoteStorageClientInfo` as an atomic unit. This
means that, when invoking the `gather` command, the user has to either
provided *no* data related to the remote storage server, or *all* data
required for the remote storage server. That is to say, the url,
username, and password. If the user provides only one or two of these
arguments, then the parsing of the arguments will fail.

A separate behavior is to allow the user to provide any piece of
information related to the remote storage server, and only require all
pieces of information in the `combineToInstructions` step.