Parsing fails on invalid configuration value
If the user provides a value for a configuration - be it via the
command-line, environment, or a configuration file - then failure to
parse the value should propogate. The parsing failure should not be
"swallowed up" as if the user failed to provide that option at all -
i.e., we should not default to a Nothing value.