Server responds with IDs of Server Entries.
The expectation is that the client will persist the IDs as well, as
proof that the entries in the client have been sync'd to the server.

Notice that in this implementation, we are returning a list of the
server IDs. The original specification was to return a list map of
client ids to server ids. In this list implementation, we are relying on
the fact that a list is an ordered collection. It is likely that a list
implementation is too prone to errors, and we may decide to use a map in
the future.