`gather` command does not send to sync server.
There is no reason for the `gather` command to send data to another
service. Instead, the user should run the `sync` command - which will be
implemented soon - in order to do this. The user can even automate this
by using `cron` or some other utility to run the sync command
periodically.

See the conversation [on
GitHub](https://github.com/NorfairKing/hastory/issues/13) for details.