POST "/entries" returns no content.
It is important for the sync server to not return any identifier when a
client sends data to the server.

The reason for this is that when the client pulls remote data down, the
client is responsible for sending to the server the client's position in
the "replication log". If the server were to send back identifiers for
the newly created entries when a client pushes data, then the following
error could occur:

Client 1 has sync'd two entries to the server, which have been assigned
the identifiers 1 and 2 by the server.

Client 2 creates one entry on the server; it is assigned the id 3 by the
server, and returned to the client. The client updates its local
database. The client now thinks its position in the replication log is
up to entry 3. When it asks the server for new data, the server doesn't
send entries 1 and 2 to client 2, because the server contains no data
past "position 3 in the replication log".

By NOT returning any identifiers when a client syncs to the server, the
scenario above cannot occur.