Update existing local entry when syncing.
When syncing remote entries to the local database, some of those entries
will already exist in the local database, but won't have their syncWitness
set. This situation arises when the entry was in fact created by the
client. In this case, we don't want to create a new entry when fetching
- instead, we want to update the existing entry.