Remove remote storage info from flags.
The `Flags` data type should be used to parse arguments that are *not*
specific to a command. The only command that will make use of the remote
storage info is the `sync` command. As a result, the `Flags` data type
should not contain remote storage information.

[Options Parsing
Reference](https://cs-syd.eu/posts/2020-05-14-option-parsing-2).